+++ lib/PhenixSdk.xcframework/ios-arm64/PhenixSdk.framework/Headers/PhenixChatMessage.h
@@ -1,0 +1,23 @@
+/**
+ * Copyright 2020 Phenix Real Time Solutions, Inc. Confidential and Proprietary. All Rights Reserved.
+ */
+#import <Foundation/Foundation.h>
+
+#import "PhenixChatUser.h"
+#import "PhenixObservable.h"
+
+@protocol PhenixChatMessage<NSObject>
+
+- (NSString*)getMessageId;
+
+- (PhenixObservable<NSDate*>*)getObservableTimeStamp;
+
+- (PhenixObservable<NSDate*>*)getObservableDeletedTimeStamp;
+
+- (PhenixObservable<id<PhenixChatUser>>*)getObservableFrom;
+
+- (PhenixObservable<NSString*>*)getObservableMessage;
+
+- (PhenixObservable<NSString*>*)getObservableMimeType;
+
+@end