+++ lib/PhenixSdk.xcframework/ios-arm64/PhenixSdk.framework/Headers/PhenixDeviceCapability.h
@@ -1,0 +1,19 @@
+/**
+ * Copyright 2020 Phenix Real Time Solutions, Inc. Confidential and Proprietary. All Rights Reserved.
+ */
+#import <Foundation/Foundation.h>
+
+typedef NS_ENUM(NSInteger, PhenixDeviceCapability) {
+  PhenixDeviceCapabilityWidth,
+  PhenixDeviceCapabilityHeight,
+  PhenixDeviceCapabilityFrameRate,
+  PhenixDeviceCapabilityFacingMode,
+  PhenixDeviceCapabilityFlashMode,
+  PhenixDeviceCapabilityDeviceId,
+  PhenixDeviceCapabilityLocation,
+  PhenixDeviceCapabilityPolarPattern,
+  PhenixDeviceCapabilityAudioEchoCancelationMode,
+  PhenixDeviceCapabilityVideoSourceRotationMode,
+  PhenixDeviceCapabilityAutoFocusMode,
+  PhenixDeviceCapabilityAudioChannelLayout
+};