+++ lib/PhenixSdk.xcframework/ios-arm64/PhenixSdk.framework/Headers/PhenixDimensions.h
@@ -1,0 +1,9 @@
+/**
+ * Copyright 2022 Phenix Real Time Solutions, Inc. Confidential and Proprietary. All Rights Reserved.
+ */
+#include <stddef.h>
+
+struct PhenixDimensions {
+  size_t width;
+  size_t height;
+};