+++ lib/PhenixSdk.xcframework/ios-arm64/PhenixSdk.framework/Headers/PhenixLocation.h
@@ -1,0 +1,7 @@
+/**
+ * Copyright 2018 PhenixP2P Inc. Confidential and Proprietary. All Rights Reserved.
+ */
+#import <Foundation/Foundation.h>
+
+// Indicates roughly where a source (e.g. microphone or camera) is mounted on the device (e.g. iPhone)
+typedef NS_ENUM(NSInteger, PhenixLocation) { PhenixLocationAutomatic, PhenixLocationUpper, PhenixLocationLower };