+++ lib/PhenixSdk.xcframework/ios-arm64/PhenixSdk.framework/Headers/PhenixOptionalAction.h
@@ -1,0 +1,14 @@
+/**
+ * Copyright 2018 PhenixP2P Inc. Confidential and Proprietary. All Rights Reserved.
+ */
+#import <Foundation/Foundation.h>
+
+@protocol PhenixOptionalAction<NSObject>
+
+- (BOOL)isPresent;
+
+- (void)perform;
+
+- (void)dismiss;
+
+@end