+++ lib/PhenixSdk.xcframework/ios-arm64/PhenixSdk.framework/Headers/PhenixRendererStatistics.h
@@ -1,0 +1,8 @@
+/**
+ * Copyright 2022 Phenix Real Time Solutions, Inc. Confidential and Proprietary. All Rights Reserved.
+ */
+#import <Foundation/Foundation.h>
+
+struct PhenixRendererStatistics {
+  NSTimeInterval lag;
+};