+++ lib/PhenixSdk.xcframework/ios-arm64_x86_64-simulator/PhenixSdk.framework/Headers/PhenixFlashMode.h
@@ -1,0 +1,10 @@
+/**
+ * Copyright 2018 PhenixP2P Inc. Confidential and Proprietary. All Rights Reserved.
+ */
+#import <Foundation/Foundation.h>
+
+typedef NS_ENUM(NSInteger, PhenixFlashMode) {
+  PhenixFlashModeAutomatic,
+  PhenixFlashModeAlwaysOn,
+  PhenixFlashModeAlwaysOff
+};