+++ lib/PhenixSdk.xcframework/ios-arm64_x86_64-simulator/PhenixSdk.framework/Headers/PhenixMemberState.h
@@ -1,0 +1,12 @@
+/**
+ * Copyright 2018 PhenixP2P Inc. Confidential and Proprietary. All Rights Reserved.
+ */
+#import <Foundation/Foundation.h>
+
+typedef NS_ENUM(NSInteger, PhenixMemberState) {
+  PhenixMemberStateActive,
+  PhenixMemberStatePassive,
+  PhenixMemberStateHandRaised,
+  PhenixMemberStateInactive,
+  PhenixMemberStateOffline
+};