+++ lib/PhenixSdk.xcframework/ios-arm64_x86_64-simulator/PhenixSdk.framework/Headers/PhenixRequestStatusObject.h
@@ -1,0 +1,14 @@
+/**
+ * Copyright 2020 Phenix Real Time Solutions, Inc. Confidential and Proprietary. All Rights Reserved.
+ */
+#import <Foundation/Foundation.h>
+
+#import "PhenixRequestStatus.h"
+
+// clang-format off
+__attribute__((visibility("default"))) @interface PhenixRequestStatusObject : NSObject
+
+@property (nonatomic) PhenixRequestStatus status;
+
+@end
+// clang-format on