+++ lib/Universal/PhenixSdk.framework/Headers/PhenixDimensions.h
@@ -1,9 +1,0 @@
-/**
- * Copyright 2022 Phenix Real Time Solutions, Inc. Confidential and Proprietary. All Rights Reserved.
- */
-#include <stddef.h>
-
-struct PhenixDimensions {
-  size_t width;
-  size_t height;
-};