+++ lib/Universal/PhenixSdk.framework/Headers/PhenixMediaType.h
@@ -1,6 +1,0 @@
-/**
- * Copyright 2018 PhenixP2P Inc. Confidential and Proprietary. All Rights Reserved.
- */
-#import <Foundation/Foundation.h>
-
-typedef NS_ENUM(NSInteger, PhenixMediaType) { PhenixMediaTypeAudio, PhenixMediaTypeVideo };