+++ lib/Universal/PhenixSdk.framework/Headers/PhenixSdk.h
@@ -1,9 +1,0 @@
-/**
- * Copyright 2018 PhenixP2P Inc. Confidential and Proprietary. All Rights Reserved.
- */
-#import "PhenixChannelExpressFactory.h"
-#import "PhenixPCastExpressFactory.h"
-#import "PhenixPCastFactory.h"
-#import "PhenixRoomChatServiceFactory.h"
-#import "PhenixRoomExpressFactory.h"
-#import "PhenixRoomServiceFactory.h"