+++ lib/iPhoneOS/PhenixSdk.framework/Headers/PhenixChannelExpressOptions.h
@@ -1,8 +1,0 @@
-/**
- * Copyright 2018 PhenixP2P Inc. Confidential and Proprietary. All Rights Reserved.
- */
-#import <Foundation/Foundation.h>
-
-@protocol PhenixChannelExpressOptions<NSObject>
-
-@end