+++ lib/iPhoneOS/PhenixSdk.framework/Headers/PhenixRendererStatistics.h
@@ -1,8 +1,0 @@
-/**
- * Copyright 2022 Phenix Real Time Solutions, Inc. Confidential and Proprietary. All Rights Reserved.
- */
-#import <Foundation/Foundation.h>
-
-struct PhenixRendererStatistics {
-  NSTimeInterval lag;
-};